build(helm/upgrade): prepare chart before packing container
As things stand the tree is always pointing to the stable helm version, eg: 2.1.0
When building upgrade images for pre-releases or patches we need to patch the chart
so that it may point to the correct version.
This is done using the same publish script that is used to publish the chart to the
helm repo.

If we're building without a tag then for now we simply ignore this..

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>