feat(helm/nvme): add more knobs to control timeouts and qdepth
Adds default timeout of 120s.
This is an exaggerated value for local disks but we've observed that in
shared virtual environments having a higher timeout value is beneficial.
Please adjust this according to your hardware and needs.
Also please note in certain cases, you may have to set this to an even
larger timeout, eg: 300s.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>