refactor(chart): provide CRDs in a subchart
This changeset is mainly intended to skip Jaeger CRDs if it is not
enabled. This is especially relevant when Jaeger tracing is installed
separately from OpenEBS Mayastor.

Helm 3 does not manage CRDs.
- https://helm.sh/docs/chart_best_practices/custom_resource_definitions

`helm uninstall` won't remove CRDs, and `helm updgrade` won't upgrade
them. Manual intervention is required with the current setup.
 - CRDs installed from `crds` folder are not included into the
 helm release.

Additionally, it is not possible to template CRDs in Helm 3 in the
`crds` folder.

Moving CRDs to a chart solves several issues: clean uninstall,
possibility for upgrades, templating.

Also make volumeSnapshots CRD optional

Signed-off-by: cmontemuino <1761056+cmontemuino@users.noreply.github.com>