feat(chart): prevent CRDs from being recycled
Problem: with current implementation a chart uninstal also removes the
CRDs, which triggers deletion of all created CRs.

Solution: add a configuration to keep CRDs by default, instructing Helm
to not delete such resources. Also add the possibility to specify extra
annotations. This is especially interesting for Argo CD users, as Argo
CD does not install charts the Helm way. In such a case, an extra
annotation to prevent prunning resources is required.

Signed-off-by: cmontemuino <1761056+cmontemuino@users.noreply.github.com>