fix(jaeger): update and use the jaeger bundled CRD
It seems otherwise there's a race condition in which CRD gets installed
depending on the helm version.
We don't actually need to use our own CRD for jaeger anymore, so let's
instead if the bundled one from jaeger-operator.
Updates jaeger-operator.
Note that as of now cert-manager is a required dependency for jaeger.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>