feat(helm): add list of container images
Adds a list of container images which are used by the chart.
This list is contained in file: charts/images.txt

Sometimes the images can't be figured out using templates, and in such
cases we need to manually add them.
charts/dependencies-images.txt is covering the this for the dependencies.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>