build: cleanup helm deps as folders
The helm release.sh downloads the deps as folders which is then used by the
build sandbox.
This however creates a buggy behaviour in the image list script, if it is
run ahead of the release script.
Here we simply clean up the folders.
todo: we don't even need them to begin with, we can use the tar files!

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>